requirements.txt
beautifulsoup4
openai
Pillow
pytesseract
requests
PySocks
opencv-python
